you can now win money!!!
i made is so the players can now win money. the windows are a bit
glitchy because of my to front method, i tried to fix it by using to
back before hand but that didnt help at all and without a to front or
set visible the whole thing goes to the task bar and you have to click
it for it to open. and there are a few glitches when guessing stuff
sometimes it just decides to stop working. again probably stupid while
loop errors